Jeremy - update paypal i18n

See merge request Lin/lss_vue_enigma!65

+++ src/views/mkt-plugin/lucky-draw/DrawList.vue
@@ -1,52 +1,52 @@
 <template>
     <div class="m-3 sm:m-5">
-        <div class="flex justify-between flex-col lg:flex-row gap-5 mb-10">
-            <div class="flex justify-start">
-                <label class="form-label mr-10">{{ $t('lucky_draw.draw_list.campaign_title') }}</label>
-                <h2 class="w-42"> {{ props.campaignTitle }} </h2>
+        <div class="flex justify-between flex-col xl:flex-row gap-5 mb-10">
+            <div class="flex justify-start font-medium">
+                <label class="form-label mr-10 my-auto">{{ $t('lucky_draw.draw_list.campaign_title') }}</label>
+                <h2 class="w-42 my-auto"> {{ props.campaignTitle }} </h2>
             </div>
-            <div class="flex justify-between">
+            <div class="flex justify-end">
+                <button class="btn btn-primary h-[35px] sm:h-[42px] my-auto mr-6 w-40" @click="toManageOrder()"> {{ $t(`campaign_live.incoming.manage_order` ) }} </button>
                 <button class="btn btn-primary w-32 mr-0 sm:mr-3 h-[35px] sm:h-[42px]" @click="eventBus.emit('showWinnersList')">
                     {{ $t('lucky_draw.draw_list.all_winner') }}
-                </button>
-                <button class="btn btn-primary w-32 ml-auto h-[35px] sm:h-[42px]" @click="existsDrawlist = eventBus.emit('changeDrawPage')">
-                    {{ $t('lucky_draw.draw_list.create') }}
                 </button>
             </div>
         </div>
         <div v-for="(luckydraw, index) in props.luckydrawList" :key="index"
             class="box bg-[#f1f1f1] relative text-left" >     
-            <div class="flex flex-row flex-wrap justify-start lg:justify-between m-[0.7rem] p-5 px-3 lg:p-5 lucky-set">
-                <div class="flex flex-col xl:flex-row justify-start w-[70px] lg:w-[120px] mr-5 my-auto"> 
-                    <img class="h-[120px] object-cover lg:mr-5" :src="storageUrl + luckydraw.prize.image" />
+            <div class="flex flex-row flex-wrap justify-start lg:justify-between m-[0.7rem] p-5 px-3 xl:p-5 lucky-set">
+                <div class="flex flex-col xl:flex-row justify-start w-[70px] md:w-[100px] lg:w-[200px] mr-5 md:mr-10 lg:mr-20 xl:mr-5 my-auto"> 
+                    <img class="h-[120px] object-cover xl:mr-5" :src="storageUrl + luckydraw.prize.image" />
                 </div>
-                <span class="h-auto w-40 break-words text-[16px] hidden lg:block">{{ luckydraw.prize.name }}</span>
-                <div class="flex flex-col lg:flex-row flex-wrap justify-start md:mt-5 w-[55%] lg:w-auto lg:ml-auto">  
-                    <span class="my-auto lg:hidden break-words w-full">{{ luckydraw.prize.name }}</span>
-
-                    <div class="lg:border-r-2 border-white flex flex-row lg:flex-col w-full lg:pr-5 lg:w-36 xl:w-44 text-sm lg:text-lg justify-between lg:justify-center text-right">
-                        <div> 
-                            <span v-if="drawTitleMap[luckydraw.type] == 'Draw Like'" class="text-slate-500 whitespace-nowrap mr-auto lg:mr-0"> {{ $t(`lucky_draw.draw_list.${luckydraw.type}`) }} </span>
-                            <span v-else-if="drawTitleMap[luckydraw.type] == 'Draw Purchased'" class="text-slate-500 whitespace-nowrap mr-auto lg:mr-0"> {{ $t(`lucky_draw.draw_list.${luckydraw.type}`) }} </span>
-                            <span v-else class="text-slate-500 order-1 whitespace-nowrap mr-0 hidden md:block"> {{ $t(`lucky_draw.draw_list.${luckydraw.type}`) }} </span>
+                <span class="h-auto w-40 break-words text-[16px] hidden 2xl:block">{{ luckydraw.prize.name }}</span>
+                <div class="flex flex-col 2xl:flex-row flex-wrap justify-start xl:mt-5 w-[55%] xl:w-auto xl:ml-auto">  
+                    <span class="my-auto 2xl:hidden break-words w-full">{{ luckydraw.prize.name }}</span>
+                    
+                    <div class="flex flex-col xl:flex-row flex-wrap" > 
+                        <div class="xl:border-r-2 border-white flex flex-row xl:flex-col w-full xl:pr-5 xl:w-36 xl:w-44 text-sm xl:text-xl justify-between xl:justify-center text-right">
+                            <div> 
+                                <span v-if="drawTitleMap[luckydraw.type] == 'Draw Like'" class="text-slate-500 whitespace-nowrap mr-auto xl:mr-0"> {{ $t(`lucky_draw.draw_list.${luckydraw.type}`) }} </span>
+                                <span v-else-if="drawTitleMap[luckydraw.type] == 'Draw Purchased'" class="text-slate-500 whitespace-nowrap mr-auto xl:mr-0"> {{ $t(`lucky_draw.draw_list.${luckydraw.type}`) }} </span>
+                                <span v-else class="text-slate-500 order-1 whitespace-nowrap mr-0 hidden md:block"> {{ $t(`lucky_draw.draw_list.${luckydraw.type}`) }} </span>
+                            </div>
+                            <div class="h-[22px] xl:h-[28px] mr-auto xl:mr-0 xl:my-1"> 
+                                <span class="text-[#E75F34]" v-if="luckydraw.type == 'product'"> ({{ luckydraw.campaign_product.order_code }}) {{ luckydraw.campaign_product.name }} </span>
+                                <span class="text-[#E75F34]" v-else-if="luckydraw.type == 'keyword'"> {{ luckydraw.comment }} </span>
+                            </div>
                         </div>
-                        <div class="h-[22px] lg:h-[28px] mr-auto lg:mr-0 lg:my-1"> 
-                            <span class="text-[#E75F34]" v-if="luckydraw.type == 'product'"> ({{ luckydraw.campaign_product.order_code }}) {{ luckydraw.campaign_product.name }} </span>
-                            <span class="text-[#E75F34]" v-else-if="luckydraw.type == 'keyword'"> {{ luckydraw.comment }} </span>
+                        <div class="xl:border-r-2 border-white flex flex-row xl:flex-col w-full xl:px-5 xl:w-36 text-sm xl:text-xl justify-between xl:justify-center">
+                            <div class="text-slate-500 mr-3"> {{ $t('lucky_draw.draw_list.winners') }} </div>
+                            <span class="xl:my-1"> {{ luckydraw.num_of_winner }} </span>
                         </div>
-                    </div>
-                    <div class="lg:border-r-2 border-white flex flex-row lg:flex-col w-full lg:px-5 lg:w-36 text-sm lg:text-lg justify-between lg:justify-center">
-                        <div class="text-slate-500 mr-3"> {{ $t('lucky_draw.draw_list.winners') }} </div>
-                        <span class="lg:my-1"> {{ luckydraw.num_of_winner }} </span>
-                    </div>
-                    <div class="flex flex-row lg:flex-col w-full lg:w-44 lg:px-5 text-sm lg:text-lg justify-between lg:justify-center">
-                        <span class="text-slate-500 mr-auto lg:mr-3 whitespace-nowrap"> {{ $t('lucky_draw.draw_list.winner_repeat') }} </span>
-                        <span v-if="luckydraw.repeatable === true" class="lg:order-1 lg:my-1"> {{ $t('lucky_draw.draw_list.yes') }} </span> 
-                        <span v-if="luckydraw.repeatable === false" class="lg:my-1 "> {{ $t('lucky_draw.draw_list.no') }} </span>
+                        <div class="flex flex-row xl:flex-col w-full xl:w-44 xl:px-5 text-sm xl:text-xl justify-between xl:justify-center">
+                            <span class="text-slate-500 mr-auto xl:mr-3 whitespace-nowrap"> {{ $t('lucky_draw.draw_list.winner_repeat') }} </span>
+                            <span v-if="luckydraw.repeatable === true" class="xl:order-1 xl:my-1"> {{ $t('lucky_draw.draw_list.yes') }} </span> 
+                            <span v-if="luckydraw.repeatable === false" class="xl:my-1 "> {{ $t('lucky_draw.draw_list.no') }} </span>
+                        </div>
                     </div>
                 </div>
-                <div class="mt-5 lg:m-0 flex w-[100%] 2xl:w-fit justify-end">
-                    <button class="btn btn-primary w-full lg:w-32 mt-auto h-[35px] sm:h-[42px] ml-auto" @click="goDraw(luckydraw.id)">
+                <div class="mt-5 xl:mt-0 flex w-[100%] xl:w-fit ml-auto">
+                    <button class="btn btn-primary w-full xl:w-32 mt-auto h-[35px] sm:h-[42px] ml-auto" @click="goDraw(luckydraw.id)">
                         {{ $t('lucky_draw.draw_list.start') }}
                     </button>
                 </div>
@@ -57,8 +57,10 @@
                 </DropdownToggle>
                 <DropdownMenu class="w-30">
                     <DropdownContent>
-                        <DropdownItem @click="editDraw(luckydraw.id)">{{ $t('lucky_draw.draw_list.edit') }}</DropdownItem>
-                        <DropdownItem @click="deleteDraw(luckydraw.id)">{{ $t('lucky_draw.draw_list.delete') }}</DropdownItem>
+                        <DropdownItem @click="editDraw(luckydraw.id)" class="w-24 text-center whitespace-nowrap text-[14px]" >
+                            <EditIcon class="w-[20px] h-[20px] mx-1"/> {{ $t('lucky_draw.draw_list.edit') }}</DropdownItem>
+                        <DropdownItem @click="deleteDraw(luckydraw.id)" class="w-24 text-center text-danger whitespace-nowrap text-[14px]">
+                            <Trash2Icon class="w-[20px] h-[20px] mx-1"/> {{ $t('lucky_draw.draw_list.delete') }}</DropdownItem> 
                     </DropdownContent>
                 </DropdownMenu>
             </Dropdown>
@@ -80,6 +82,7 @@ import youtube_platform from '/src/assets/images/lss-img/youtube.png';
 import facebook_platform from '/src/assets/images/lss-img/facebook.png';
 import instagram_platform from '/src/assets/images/lss-img/instagram.png';
 import unbound from '/src/assets/images/lss-img/noname.png';
+import i18n from "@/locales/i18n"
 
 const props = defineProps({
     luckydrawList: Object,
@@ -96,6 +99,9 @@ const drawTitleMap = ref({
     keyword: "Draw Keyword" 
 })
 
+const toManageOrder = ()=>{
+    router.push({ name: 'manage-order', params: { campaign_id: route.params.campaign_id}})
+}
 
 const goDraw = (lucky_draw_id) => {
     let routeData = router.resolve({ name: 'lucky-draw-flow', params: {lucky_draw_id: lucky_draw_id} })
@@ -110,12 +116,17 @@ const editDraw = (lucky_draw_id) => {
 
 const deleteDraw = (lucky_draw_id) => {
     hideDropDown()
-    delete_campaign_lucky_draw(lucky_draw_id).then(res => {
+    let yes = confirm(`${i18n.global.t("lucky_draw.draw_list.confirm_delete")}`)
+    if(yes){
+        delete_campaign_lucky_draw(lucky_draw_id).then(res => {
         router.go()
-    }).catch(err => {
-        console.log(err)
-    })
+        }).catch(err => {
+            console.log(err)
+        })
+    }
 }
+
+
 const hideDropDown = ()=>{
   dom('.dropdown-menu').removeClass('show')
 }
